fix: DB 커넥션 풀을 사용하지 않고 단일 커넥션 연결로 변경
